requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
pyyaml==6.0.1
telethon==1.28.5
g4f==0.1.9.0
aiohttp==3.8.5
asyncio==3.4.3
cryptg==0.4.0
pillow==10.0.0
python-dateutil==2.8.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3